refactor: mode Field/Game/Room/Player/Ship to model folder
